Regroupement des fonctions dans outils.php / utilisation de require_once
