+++ ajouter_avis.php
@@ -3,29 +3,7 @@
 global $dbh;
 session_start();
 include('connect.php');
-
-// Fonction pour ajouter un avis sur un produit
-function ajouterAvis($dbh, $product_id, $user_id, $note, $commentaire) {
-    // Vérification si la note est valide (entre 1 et 5)
-    if ($note >= 1 && $note <= 5) {
-        $sql = "INSERT INTO evaluations (produit_id, utilisateur_id, note, commentaire) VALUES (:product_id, :user_id, :note, :commentaire)";
-        $stmt = $dbh->prepare($sql);
-
-        $stmt->bindParam(":product_id", $product_id, PDO::PARAM_INT);
-        $stmt->bindParam(":user_id", $user_id, PDO::PARAM_INT);
-        $stmt->bindParam(":note", $note, PDO::PARAM_INT);
-        $stmt->bindParam(":commentaire", $commentaire, PDO::PARAM_STR);
-
-        // Exécution de la requête d'insertion de l'avis
-        if ($stmt->execute()) {
-            return true; // Succès : avis ajouté avec succès
-        } else {
-            return false; // Échec : erreur lors de l'ajout de l'avis
-        }
-    } else {
-        return false; // Échec : la note n'est pas valide
-    }
-}
+require_once ('./inc/outils.php');
 
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
     if (isset($_SESSION["client_id"])) {
