+++ categorie.php
@@ -2,50 +2,7 @@
 // Inclusion du fichier de connexion à la base de données
 global $dbh;
 require 'connect.php';
-
-// Fonction pour récupérer le nom de la catégorie
-function getNomCategorie($dbh, $cat_id) {
-    try {
-        $stmtCat = $dbh->prepare("SELECT nom FROM categories WHERE id = :cat_id");
-        $stmtCat->bindParam(':cat_id', $cat_id, PDO::PARAM_INT);
-        $stmtCat->execute();
-
-        if ($stmtCat->rowCount() > 0) {
-            return $stmtCat->fetch(PDO::FETCH_ASSOC)['nom'];
-        }
-    } catch (PDOException $e) {
-        echo "Erreur lors de la récupération du nom de la catégorie : " . $e->getMessage();
-    }
-    return '';
-}
-
-// Fonction pour récupérer la liste des produits d'une catégorie
-function getProduitsByCategorie($dbh, $cat_id) {
-    try {
-        $stmt = $dbh->prepare("SELECT p.*, 
-                                      pr.pourcentage_remise,
-                                      CASE 
-                                          WHEN pr.pourcentage_remise IS NOT NULL THEN p.prix - (p.prix * pr.pourcentage_remise / 100)
-                                          ELSE p.prix
-                                      END AS prix_apres_remise
-                               FROM produits p
-                               LEFT JOIN promotions pr ON p.id = pr.produit_id AND CURRENT_TIMESTAMP BETWEEN pr.date_debut AND pr.date_fin
-                               WHERE p.categorie_id = :cat_id");
-        $stmt->bindParam(':cat_id', $cat_id, PDO::PARAM_INT);
-        $stmt->execute();
-
-        $products = array();
-
-        while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
-            $products[] = $row;
-        }
-
-        return $products;
-    } catch (PDOException $e) {
-        echo "Erreur lors de l'exécution de la requête : " . $e->getMessage();
-    }
-    return array();
-}
+require_once ('./inc/outils.php');
 
 $cat_id = isset($_GET['cat_id']) ? $_GET['cat_id'] : 0;
