+++ editer_client.php
@@ -5,7 +5,7 @@ session_start();
 
 // Inclusion du fichier de connexion à la base de données
 include('connect.php');
-include('inc/outils.php');
+require_once ('./inc/outils.php');
 
 // Récupération de l'identifiant du client à partir de la session
 $client_id = isset($_SESSION['client_id']) ? $_SESSION['client_id'] : null;
@@ -18,10 +18,7 @@ if (!$client_id) {
 }
 
 // Récupération des informations actuelles du client depuis la base de données
-$stmtClient = $dbh->prepare("SELECT * FROM clients WHERE id = :client_id");
-$stmtClient->bindParam(':client_id', $client_id, PDO::PARAM_INT);
-$stmtClient->execute();
-$clientInfo = $stmtClient->fetch(PDO::FETCH_ASSOC);
+$clientInfo = getClientInfoEditerClient($dbh, $client_id);
 
 // Traitement du formulaire de mise à jour
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
@@ -41,46 +38,37 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         $mot_de_passe_hache = password_hash($nouveau_mot_de_passe, PASSWORD_DEFAULT);
 
         // Mettre à jour les informations du client avec le nouveau mot de passe
-        $stmtUpdate = $dbh->prepare("UPDATE clients SET
-            nom = :nom,
-            alias = :alias,
-            email = :email,
-            adresse = :adresse,
-            ville = :ville,
-            code_postal = :code_postal,
-            telephone = :telephone,
-            mot_de_passe = :mot_de_passe
-            WHERE id = :client_id");
-
-        $stmtUpdate->bindParam(':mot_de_passe', $mot_de_passe_hache, PDO::PARAM_STR);
+        $data = [
+            'nom' => $nom,
+            'alias' => $alias,
+            'email' => $email,
+            'adresse' => $adresse,
+            'ville' => $ville,
+            'code_postal' => $code_postal,
+            'telephone' => $telephone,
+            'mot_de_passe' => $mot_de_passe_hache,
+        ];
     } else {
         // Mettre à jour les informations du client sans changer le mot de passe
-        $stmtUpdate = $dbh->prepare("UPDATE clients SET
-            nom = :nom,
-            alias = :alias,
-            email = :email,
-            adresse = :adresse,
-            ville = :ville,
-            code_postal = :code_postal,
-            telephone = :telephone
-            WHERE id = :client_id");
+        $data = [
+            'nom' => $nom,
+            'alias' => $alias,
+            'email' => $email,
+            'adresse' => $adresse,
+            'ville' => $ville,
+            'code_postal' => $code_postal,
+            'telephone' => $telephone,
+            'mot_de_passe' => $clientInfo['mot_de_passe'], // Réutilisation de l'ancien mot de passe
+        ];
     }
 
-    $stmtUpdate->bindParam(':nom', $nom, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':alias', $alias, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':email', $email, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':adresse', $adresse, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':ville', $ville, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':code_postal', $code_postal, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':telephone', $telephone, PDO::PARAM_STR);
-    $stmtUpdate->bindParam(':client_id', $client_id, PDO::PARAM_INT);
-
-    if ($stmtUpdate->execute()) {
+    // Appel de la fonction pour mettre à jour les informations du client
+    if (miseAJourClientInfo($dbh, $client_id, $data)) {
         // Rediriger vers la page compte_client.php après la mise à jour
         header('Location: compte_client.php');
         exit();
     } else {
-        echo "Erreur lors de la mise à jour des informations du client : " . $stmtUpdate->errorInfo()[2];
+        echo "Erreur lors de la mise à jour des informations du client.";
     }
 }
 ?>
