+++ expedition.php
@@ -2,6 +2,7 @@
 global $dbh;
 session_start();
 include('connect.php');
+require_once ('./inc/outils.php');
 
 if (!isset($_SESSION['email'])) {
     header('Location: connexion_inscription.php');
