+++ head_header_nav.php
@@ -4,6 +4,7 @@ if (session_status() == PHP_SESSION_NONE) {
     session_start();
 }
 require 'connect.php';
+require_once ('./inc/outils.php');
 error_reporting(E_ALL);
 ini_set('display_errors', '1');
 echo '<!DOCTYPE html>
